atomupd: Reconfigure the current bootloader after upgrade
The GRUB configuration file of the running system (i.e. self) should be
updated since the bootloader configuration of GRUB displays the entry of
the duplicated system in its menu (i.e. other).

This commit reconfigures the current configuration of GRUB after the
update of the redundant system is completed (i.e. right after its
bootloader is setup).