dracut: Mount /etc even if roothash is not specified
The mountpoint /etc is mounted only if the kernel argument roothash is
not set.

This leads to different behaviours if verity is disabled (i.e. roothash
is not specified).

For example, if the hostname is modified and the new name set in the
/etc overlay, then that name is set by the service systemd-hostnamed.
The /etc overlay is mounted latter anyway by SteamOS.

The hostname is updated during the starting of X and changes the X
cookie. The update happens between the Display Manager and the Desktop
Environment that prevent the UI to start properly and it needs to be
restarted.

The hostname is set by systemd right after the root is switched in the
initrd.

This mounts the /etc overlay even if the kernel argument roothash is not
specified to reduce the differences wheter the verity is enabled or not.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>