glx: steamos-glx-driver: Omit the nouveau driver
It appears that dracut ignores the file in /etc/modprobe.d, Therefore,
the module nouveau is never blacklisted and it is loaded instead of the
proprietary module nvidia-dkms.

The module nouveau can be blacklisted using rd.driver.blacklist=nouveau
but this requires to sed the file /etc/default/grub. The kernel options
in /etc/cmdline.d/*.conf do not work as well.

This commit should be reverted for a proper solution.

Note: See commit 5470efa (steamos-glx-driver: Fix sed failure on read-only
filesystem).

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>