Revert "glx: steamos-glx-driver: Omit the nouveau driver"
This reverts commit a6fcaa85a2de1210fbf021a6cc3389e4c6c82a1b.

The Debian GRUB Quick-Boot feature saves the variable recordfail to the
file the grubenv that is write back to a filesystem. That filesystem is
flagged read-only, but the GRUB helper script does not care about it and
writes back the changes to the filesystem.

That filesystem is controlled by the dm-verity integrity checking, and
because it is altered by GRUB, the integrity fails while the system
installs the boot artifacts to a safe writable place.

As a consequence, the GRUB configuration is not updated and this new
boot location is never used. Therefore, the system keeps using the
factory boot artifacts that does not include the nvdia driver.