dracut: Add pre-pivot hook to mount the var partition
The /var partition is mounted using the file /etc/fstab from the new
rootfs and that file lives in the /etc overlay directory.

That overlay directory is not yet mounted as the artifacts live in that
partition and thus need the /var partition to be mounted.

As a consequence, dracut never uses the latest version of the fstab that
is confusing.

This adds a new pre-pivot hook that mounts the /var partition without
using the file /etc/fstab from the new rootfs.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>