grub: grub-install: Be less smart
On a read-only system, the grub-install option --boot-directory
/var/boot has to be set if the directory /var/boot is not bind-mounted
over /boot.

The helper script steamos-grub-install intends to be smart and adds the
options --boot-directory if the system has the kernel option roothash=
in its command line and if the directory /boot is a mountpoint.

But these too condition are too smart in the case of an atomic update is
triggered from a writable system where the argument roothash= is not
set.

In this situation the RAUC postinstallation script fails with the error
below because the kernel option roothash= is unset that breaks the magic
of auto-adding the option --boot-directory /var/boot (even if the
directory /boot is not a mount-point).

This removes the check of the kernel option roothash= and relies only if
the directory /boot is a mountpoint as in the script update-grub (that
detects the path to the grub.cfg).

Fixes:

	Installing for x86_64-efi platform.
	/usr/lib/grub/grub-install: error: cannot delete `/boot/grub/x86_64-efi/pbkdf2.mod': Read-only file system.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>