Fix "grub: grub-install: Be less smart"
This fixes commit 6808938fbe70bf1c2defe7ef77f51905690a8438.

A way to bypath the magic and to install GRUB on the real /boot
directory is *NEEDED* (i.e. running from a writable system, or from a
chroot'ed environment...); and error saying that the filesystem is
read-only is preferred:

	Installing for x86_64-efi platform.
	/usr/lib/grub/grub-install: error: cannot delete `/boot/grub/x86_64-efi/pbkdf2.mod': Read-only file system.

Under certain circumstances, the GRUB binary in the EFI partition would
not find its modules; because there is no way to install the modules on
the real root partition and the directory /var/boot may be gone with the
disabling of verity.

This moves the magic to the script steamos-bootloaders-instal where the
initial issue comes. The commit has introduced a change that replaced
call of grub-mkimage by grub-install. See the commit eaf0139 (misc:
Introduce steamos-bootloaders-install).

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>