dracut: Be more friendly with other distros
The couple GRUB/osprober of standard distros detects SteamOS as a
bootable distro; therefore a menu is added to GRUB. os-prober for
SteamOS is patched but they are not yet upstreamed.

Unfortunatly, the other distros cannot boot the initial ramdisk for
SteamOS as is. The utility linux-boot-prober that is shipped along with
os-prober intends to be smart by looking for kernel and initial ramdisk
in the disks and try to guess the kernel command-line.

But linux-boot-prober did not have the knowledge for the partition
scheme (efi, verity, root, var, home) and the need for the *MANDATORY*
argument steamos.efi=BLOCKDEV that is specific to initial ramdisk of
SteamOS.

This argument tells to the SteamOS's dracut module where to find the
pseudo-EFI partition that contains the set of the partitions.

This set contains /var partition that *MUST* be mounted to provide a
writable space to copy the modules and to create some symlinks with the
rootfs. Without these symlinks, the kernel *CANNOT* find its modules and
breaks the boot of SteamOS.

This installs the files to the initial ramdisk so the dracut module can
fallback to that file if the arguement steamos.efi=BLOCKDEV is omitted.

Note: This is a temporary solution as this tainted the initial ramdisk
with data that are specific to the machine. Another solution would
consist in finding the var partition using the GPT Partition Name of the
root partition (i.e. PARTLABEL) and the prefix -A/-B as SteamOS relies
on the GPT Partition Names.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>