steamos-verity: Remake GRUB image after enable
TL;DR; this re-builds the GRUB image after verity is re-enabled to fix a
boot issue.

The boot is broken because GRUB cannot find its modules after verity is
re-enabled. See the log below:

>
>                              GNU GRUB  version 2.04
>
>  ����������������������������������������������������������������������������Ŀ
>  �*SteamOS                                                                    �
>  � Advanced options for SteamOS                                               �
>  � System setup                                                               �
>  � System halt                                                                �
>  � System restart                                                             �
>  �                                                                            �
>  �                                                                            �
>  �                                                                            �
>  �                                                                            �
>  �                                                                            �
>  �                                                                            �
>  �                                                                            �
>  ������������������������������������������������������������������������������
>
>       Use the ^ and v keys to select which entry is highlighted.
>       Press enter to boot the selected OS, `e' to edit the commands
>       before booting or `c' for a command-line. ESC to return previous
>       menu.
>
> (...)
>
> error: file `/boot/grub/x86_64-efi/all_video.mod' not found.
> error: file `/boot/grub/x86_64-efi/steamenv.mod' not found.
> error: file `/boot/grub/x86_64-efi/gzio.mod' not found.
> error: can't find command `steamenv_boot'.
> error: can't find command `initrd'.
>
> Press any key to continue...

All the entries in the menu are broken and the only alternative is to
enter the command-line by pressing `c'.

GRUB looks for its modules using the variable $prefix which is set to
`(hd0,gpt5)/boot/grub` (aka the `/boot/grub` directory on the /var
partition). That directory is removed though to clear all former `/boot`
artifacts.

>
>                              GNU GRUB  version 2.04
>
>    Minimal BASH-like line editing is supported. For the first word, TAB
>    lists possible command completions. Anywhere else TAB lists possible
>    device or file completions. ESC at any time exits.
>
>
> grub> echo $prefix
> (hd0,gpt5)/boot/grub

The only way to boot is to tell to GRUB to look for its modules at the
only place where they remain: in the root partition.

Set the $prefix variable to `(hd0,gpt4)/boot/grub` (aka the `/boot/grub`
directory on the / partition) and reload the configuration file then
(i.e. `(hd0,gpt2)/efi/steamos/grub.cfg`).

> grub> prefix=(hd0,gpt4)/boot/grub
> grub> configfile (hd0,gpt2)/efi/steamos/grub.cfg

The GRUB image in `/efi/EFI/steamos/grubx64.efi` contains the $prefix;
it can be overridden by the early/bootstrap config, better to update the
image though.

Therefore, the GRUB image *MUST* be rebuilt after the `/boot` directory
is removed from the /var partition.