grub: Apply the prefix to grub early config
The GRUB utility grub-mkimage implies --prefix=(memdisk)/boot/grub if
--memdisk=FILE is set; therefore the --prefix that may have been passed
is lost.

This applies the prefix directly to the grub early config to avoid to
lost for the prefix.

Note: It uses --target=compatibility_hint instead of efi_hint because
the last appends a trailing space. This is the same target that is used
by the helper prepare_grub_to_access_device in grub-mkconfig_lib.