dracut: Export steamos.efi= boot param
The couple GRUB+os-prober on other distributions (Debian, Fedora...)
detects SteamOS initrd and kernel and auto guess the command-line based
on the rootfs device on which SteamOS is installed.

The root= parameter is not enough to boot SteamOS as the initrd requires
the argument steamos.efi= to be set to look for the PARTUUID of the /var
partition to mount during early boot (because of modules are diverted to
/var/lib).

Therefore, as dracut automagically guess the root= argument, it should
complete the command-line as well with the missing steamos.efi= argument
to let GRUB+os-prober of other distribution boots SteamOS.

Warning: this works on the none-fallback initrd image only and if the
parameter --no-hostonly-cmdline is not set while the initrd image is
generated. Embedding the steamos.efi= parameter is hacky and it is added
for parity feature with root=. A better solution consists in creating a
steamos-generator for partsets that would somehow creates the required
.mount services.

This exports the steamos.efi= boot parameter for hostonly-cmdline case
only.