dracut: Requires binary udevadm
The module steamos creates the udev rules that creates the symlinks to
/dev/disk/by-partsets and reload them using udevadm.

This makes sure that the binary udevadm is part of the initrd image.

Note: The module 95udev-rules embeds the binary udevadm already in the
initrd image; this module is a dependency of 99base. This change is for
robustness purpose; and prevent from upstream changes in dracut.