update-grub: Set GRUB_DISABLE_LINUX_PARTUUID=false if unset
The GRUB script /etc/grub.d/10_linux disables the use of PARTUUID for
compatibily with older kernels if the GRUB_DISABLE_LINUX_PARTUUID is
left unset.

This sets and exports the variable GRUB_DISABLE_LINUX_PARTUUID to false
if it is is not set in order to bypass 10_linux as SteamOS strongly
relies on PARTUUIDs, before filesystems UUIDs.

[1]: https://gitlab.steamos.cloud/packaging/grub2/-/blob/steamos/debian+arch+steam-2.04-15.3/util/grub.d/10_linux.in#L67-69