dracut: Add mount hook to mount the rootfs partition
The rootfs partition is mounted using the kernel parameter root= or
using systemd-gpt-auto-generator is that kernel parameter is unset.

The systemd-gpt-auto-generator picks the first partition with the GPT
partition type 4f68bce3-e8cd-4db1-96e7-fbcaf984b709 for x86_64 machines.
This will always be rootfs-A even if steamos.efi= refers to the efi-B.
Evern worse, the first partition can be a non SteamOS system.

The root= is specified for now, but it is probable that this kernel
parameter is left unset in a near future to let the initial ramdisk pick
the proper one.

This adds a new mount hook that mounts the rootfs partition based on the
block device /dev/disk/by-partsets/self/rootfs which is a symlink
created by the pre-mount hook steamos-udev-rules.sh.