steamos-update-os-notifier.action: Track online updates
The atomic updater updates the other partition set based on the version
of the current running system (i.e. the self partition set). It assumes
the current running system is the more recent system; also the other
system can be empty as the installer makes only one copy of the system.

Once the updates is completed the state of having the other system been
updated is not maintained by the kupdate-notifier; the application can
be restarted or whatever, and this is the duty of the update backend.

The atomic updater is not a daemon and thus the other-updated state is
lost and has to be maintained by the action script itself.

The other system cannot be mounted to check for its version as it may be
empty or can be broken. Therefore, the action script must track that
information.

This tracks the state of the other-updated system by maintaining files
in the /run filesystem.