steamos-update-os-notifier.action: Remove update-other flag
The action script prompts for reboot using DBus and the update-other
flag have to be removed if the user has cancelled the reboot.

There is no way to get what the user has chosen (ok or cancel; or even
sleep, restart, shutdown and logout).

This implements a workaround to remove the update-other flag after the
30 seconds of prompt timeout[1] plus an additionnal time of 5 seconds.

[1]: https://github.com/KDE/plasma-workspace/commit/6398d842d0762d9d7b4175f4938ab5a3a4f50b2b#diff-eafd1f8a9771c7da5dd838f1f5fb86a8794edc839d0d1c269309203c466c92d5R48