rauc: Implement custom bootloader backend
The RAUC service does not updates bundles anymore since the root= kernel
parameter is unset in the command-line.

The service guesses the actual boot slot using the kernel command-line
and ends with the following error:

	(root@steamos ~)# rauc install http://images.steamos.cloud/steamos-holo/disk.raucb --debug --override-boot-slot=self
	rauc-Message: 15:28:34.734: Debug log domains: 'rauc'
	(rauc:1904): rauc-DEBUG: 15:28:34.738: install started
	(rauc:1904): rauc-DEBUG: 15:28:34.738: input bundle: http://images.steamos.cloud/steamos-holo/disk.raucb
	(rauc:1904): rauc-DEBUG: 15:28:34.771: Trying to contact rauc service
	installing
	  0% Installing
	  0% Determining slot states
	 20% Determining slot states failed.
	100% Installing failed.
	LastError: Bootname or device of booted slot not found
	Installing `http://images.steamos.cloud/steamos-holo/disk.raucb` failed

The RAUC project handles A/B (seamless) system updates with the notion
of slots and some integrated backends (grub, efi, barebox, u-boot...).

However, despite the fact that SteamOS runs GRUB, it has its own way of
defining the A/B partitions which implement by the steam-chainloader
that is too specific to be upstreamed.

Furthermore, SteamOS hides the partitions A/B behind the devices self
and other under the directory /dev/disk/by-partsets for the sake of
simplicy in custom scripts.

Using the GRUB backend breaks the notition on self-contained partitions.
However, RAUC allows ones to creates its custom bootloader backend[1].

The implementation is quite straightforward has the script trusts on
some magic that is already there. The primary slot is always self and is
always in good state as it is currenly booted system.

This implements the SteamOS backend (steamcl and the partsets) using the
custom script backend and gets rid of the A/B in favour of self/other.
Also, the booted slot is specified to the RAUC service command-line has
by the option --override-boot-slot.

[1]: https://rauc.readthedocs.io/en/latest/integration.html#custom-bootloader-backend-interface