dracut: module-setup.sh: Disable systemd gpt-auto
The partition rootfs-A is always mounted by the systemd-gpt-generator
since the root= kernel parameter is not set anymore as both rootfs-A and
rootfs-B have the GPT partition type set to Root Partition (x86-64) (aka
4f68bce3-e8cd-4db1-96e7-fbcaf984b709).

This disables the discovering and mounting of the system by the systemd
gpt-auto generator by forcing the kernel parameter systemd.gpt_auto to
no. The mounting is performed by the dracut modules only using the
partsets defined in the pseudo EFI partition.