grub: Updates contition to ConditionPathIsReadWrite
The kernel parameter ro is not set anymore. It is not necessarly set and
it is confusing because the parameters ro and rw exist and the condition
*must* handle both parameters anyway.

Therefore, both services boot.mount and steamos-mkvarboot.services are
not started anymore; the condition needs to be updated.

systemd has the following condition[1]:

	ConditionPathIsReadWrite=

	ConditionPathIsReadWrite= is similar to ConditionPathExists= but
	verifies that the underlying file system is readable and
	writable (i.e. not mounted read-only).

This updates the unit condition to ConditionPathIsReadWrite=!/ instead
of relying on a kernel parameter that is not necessarly set.

[1]: https://www.freedesktop.org/software/systemd/man/systemd.unit.html#ConditionPathIsReadWrite=