update-grub: Disable other-OS probing when generating grub config
Other OS probing can add a lot of boot entries, including for the
_other_ steamos image when we're A/B booting.

We don't want this as it's not what we're using grub for: Our design
calls for a mostly invisible and straightforward second stage
bootloader that does a little bit of meta data management for us
on boot.

It specifically _shouldn't_ know about any other OSes that are
present: That belongs to the first stage bootloader and/or the
EFI boot menu in the firmware.