update-grub: remove questionable and unused -o arg
Having looked through steamos-customisations, steamos-image-recipes and
the complete Holo group - the -o argument is unused.

This is not at all surprising, since it does nothing useful. On the
contrary - the argument is _ignored_, a somewhat informative message
is printed, yet update-grub returns success.

As such, it is actively hostile to any form of scripting usage.

Just remove the argument in the script and the completion.
A follow-up commit will add same args parsing and explicitly error out,
when -o is provided.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>