grub install service should be start if any required files are missing
Previously this was triggered if the /boot/grub directory was missing.

It now starts if either grub config file (/efi or /boot) is missing,
or if the efi grub executable is missing.

This is necessary to support the factory reset workflow which differs
from the fresh install flow in that the /boot/grub directory already
exists.