Move factory reset handling into dracut's pre-pivot from shutdown
This changeset does the following:

Reset process is now triggered in the bootup pre-pivot phase rather
than the shutdown phase.

Factory reset stampfile moved to /esp rather than /run

Stampfile now lists all user data partitions found in the
/dev/disk/by-partsets/all directory instead of hardwiring them.

Stampfile persists if the reset process fails, so a reboot
will result in a repeated reset attempt.

Stampfile removed on success.

The goal is to move the reset process to the bootup phase
so users are less likely to interrupt it by leaning on the power
button, and to make the factory reeset repeat if it is interrupted
before it completes.