offload: add /var/lib/systemd/coredump offloading
Coredumps can take notable space, so let's offload them instead of
trying to squeeze them into the small, fixed-size /var.

In order for them to be useful, one needs to know the libraries loaded -
the OS version, hence libraries can be fetched via the journal logs.

Although we might need to make that easier/more accessible in the
future.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>