rauc: Switch to Desync
When we extract an image with RAUC, we now use Desync instead of Casync.

The error retry option has been provided in-line, instead of using the
`/etc/desync/config.json` file, because that configuration is currently
used only to handle the HTTP authentication.
Given that there is a discussion about disabling the HTTP authentication
altogether, if that happens we might be able to remove the need of this
additional Desync configuration file.
More info here https://gitlab.steamos.cloud/jupiter/tasks/-/issues/319

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>