Fixes for upgrade issues following bootconf changes
In steamos-boot-install, implement fallback logic to determine the
current partset name. Otherwise this creates an unbootable /esp
Workaround for https://gitlab.steamos.cloud/jupiter/tasks/-/issues/382

In subsequent upgrades, in rauc post-install, clear the image-invalid flag.
Otherwise boot is unable to proceed in the other partition.
Workaround for https://gitlab.steamos.cloud/jupiter/tasks/-/issues/384

Remember, the rauc post-install script that runs is in the booted OS,
not the target.