rauc: Skip invalid seeds instead of aborting the installation
In theory we could generate the rootfs seed index and, in a subsequent
occasion, use it to apply an update.

If between these two operations the user edited the rootfs, e.g.
modified some files with the developer RW mode enabled, the seed will be
considered invalid because some of its indexes will point to unexpected
block of memory.

In the rare occasions where this happens, we should still try to
complete the update.
With the `--skip-invalid-seeds` option, Desync will automatically
discard the seed, if it notice that at least one of the expected chunk
is invalid.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>