Create steamos-finalize-install
This script incorporates the old steamos-boot-install script, and
adds the network conversion needed to get existing profiles working
in a new iwd system.

The new migration steps will not run if they are not in a chroot
(unless -f is passed). Also this step can be skipped with --no-migration

steamos-boot-install is kept for backwards compatibility with old installations;
however, rauc post-install should call steamos-finalize-install instead.