Resolve boot race condition
Since the chainloader tells us the partuuid for the chosen device,
we could get asked to mount the EFI partition too soon.

Block on the efi by-partuuid link being available before trying to
pass it to mount as a parameter. This resolves a fatal boot error
that was coming up on a test unit, and may resolve some of the mysterious
boot errors in general.

Note that we drop to a dracut rescue shell but it is hidden by default
by the seamless boot stuff.