Check the UUID of an update in the pre-install hook
If we try to install an update that has the same UUID we should abort
the operation as early as possible, because that might break the system.

If the update doesn't include its UUID there isn't much that we can do,
we just continue and hope for the best.

Addresses: https://gitlab.steamos.cloud/jupiter/tasks/-/issues/435

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>