misc: add CFS tunings (sysctl + debugfs)
There are some CFS settings that seems to improve the gaming experience;
most of these tunings are considered debugfs material [0], so we can't
rely on sysctl settings. There is one tuning that is a sysctl though.

This commit introduces a systemd service + libexec counter-part to
perform the debugfs settings, also a sysctl config file to set up
the single sysctl.

[0] https://git.kernel.org/linus/8a99b6833c88

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>