desync: Set an error retry base interval of 1 second
When an HTTP request fails, Desync retries after waiting the base
interval multiplied by the number of attempts.

By default the interval is set to zero. This means that on error, Desync
will retry multiple times without waiting.

However, when we suspend and resume or when the connection is unstable,
it is desirable to wait a certain amount of time before retrying and
giving up.

`error-retry-base-interval` has been set to 1000000000 nanoseconds, aka
1 second.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>