atomic-update: Don't choose the "other" slot until the update is applied
When applying an update, rauc will set the state of the target image as
"bad", where we essentially set its mode to "reboot-other" and we set
the image as "invalid".

This is a security feature, because an update may fail to be applied,
and in that case we don't want to boot the "other" slot with a half
written image inside.

However in our current bootloader model, the primary partition
configuration is what matters, and only that will be evaluated. This is
a feature to avoid looping between the two partitions in case both were
configured as "reboot-other".
Because of that, if the primary partition has a mode of "reboot-other",
the "other" slot will always be chosen after the next reboot, even if we
set that the image was "invalid".

With this commit, before starting to apply the update, we set the
primary partition mode as "reboot". In this way, even if the update
fails, we can ensure that the current active slot will still remain the
primary.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>