Implement factory-reset-from bootloader via kernel command line
A kernel command line parameter (steamos.factory-reset=1) can
now trigger a factory reset.

This is in addition to the existing reset triggered by the presence
of reset config files in the well-known location on /esp.

The regeneration of the stage 2 bootloader config on /efi has also
been re-implemented now that we know how to make it work.