steamos-finalize - mitigation for a rare factory update bug
If a certain fs error/crash occurs in the update path, we are left
with an empty A.conf in the ESP partition. This is not updated by the
factory steamos-bootconf in the RAUC post-install. Result is that we
reboot into the same OOBE, in a mode I can only call Sisyphean.

Write a comment to the new config that at least points out the build
of the new image, and which should resolve this odd day-one failure.