atomic-update: Ensure that the rootfs is in read-only mode
A user could still re-enable the RW mode while the update is in
progress. However this is not a severe problem.
Actively writing to the rootfs while applying an update could result in
an aborted update (Desync) or an update that requires to download more
chunks than what was initially expected (Casync).

We are not re-enabling RW mode, after the update is completed, because
we assume that the user will reboot ASAP, and this current partition
will not be used anymore.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>