atomic-update: Do not copy RAUC config changes
If the user changed the RAUC configuration, we will end up carrying
those changes instead of using the config shipped with the image update.

Because the atomic update is a critical component, we just drop any
eventual user changes and always start from the clean vanilla RAUC
configuration included in the update.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>