rauc: Store the applied update buildid in a temporary file
When an update has been successfully installed, we store its buildid in
the RAUC temporary directory.

With this file we signal that an update has been installed and that a
reboot is currently pending.
This info is useful for the update scripts because it allows them to
always recover the current state, even if they crashed or got restarted.

By storing this file in `/run/rauc/`, it will be automatically deleted
after a reboot.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>