atomic-update: Store the new rootfs seed index in "/var"
After we successfully install an update, we can store it's rootfs seed
index so that we will not need to regenerate it while searching, and
applying, a subsequent update.

If for whatever reason we are not able to copy that file, we just print
a warning and try to continue.
"steamos-atomupd-client" will gracefully handle the situation where this
file is not available and it will simply generate it on the spot.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>