Port our dracut initramfs scripts to mkinitcpio
Some important differences:

 - we're on ash, not bash, in mkinitcpio's init
 - the root dir is in a different place, pre-pivot
 - iconv needs special handling
   - mkinitcpio didn't pick up its support files mby default
 - mkinitcpio has a different set of hook support functions to dracut
 - mkinitcpio only has 3 phases, not the 5 or so that dracut has
   - still enough for our purposes, but some code needs to move arouns

No attempt has been made to slim the initramfs down (yet) in this
config but a slimmed config produced on the deck itself worked.

The next step will be to add more config so that a preset config
generated during image build on a VM will produce a deck-specific
slimmed initrd image.