rauc: Store updated buildid in a more stable place
After a successful update we were storing a text file with the OS
buildid in `/run/rauc/`. This directory is created by the RAUC service
because we launch it with the option `--mount=/run/rauc`.

While this works fine in most cases, it becomes problematic when the
RAUC service crashes, or is stopped/restarted, because we will lose the
directory content.
For example, at the moment there isn't a reliable way to check if RAUC
is unexpectedly installing, in background, an update. So the common
approach to ensure that nothing is being installed is to stop or restart
the RAUC service.

For this reason we should instead place the temporary buildid file in
`/run/steamos-atomupd/` that is guaranteed to be around until a reboot
is performed.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>