dracut: remove chainloader logic
A while ago we added a magical ChainedLoader efivar (as opposed to the
existing ChainLoader ones). It served as a workaround to set the
steamos.efi argument on the kernel command line.

With that reinstated in grub/steamenv, we no longer need this clutch.

As a result we can remove a bunch of no-longer used programs from the
initrd.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>