steamos-ls: remove no-longer used tool
The steamos-ls tool was used exclusively to parse its output for the
now-removed steamos-parse-loaders.sh.

With the dracut script gone, we no longer need this tool.

Note: this is a fallback of a fallback path, which is broken. In
particular, this introduces another round of bootlogic and does not
consider the use-selection via the boot menu.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>