Add static server configuration
Older versions of `steamos-atomupd`, like the one that we currently have
in Jupiter rel `20221005.1`, just understand QueryUrl and ignores all
the other key-values in the configuration.

Newer versions of `steamos-atomupd` (since `jupiter-20220310`), will use
the MetaUrl value, if available, while keeping the QueryUrl as a
fallback.

MetaUrl is for the new static server, that is expected to eventually
replace the older QueryUrl.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>