misc: Reduce dracut output pollution
After the movement to btrfs as the root filesystem, due to some
incompatibility with extended attributes, dracut dumps >200K of
harmless yet alarming messages during the initramfs creation.

Not that a regular user creates initrams in a daily basis, but
this is quite annoying for developers or any user installing
their own kernels. Let's get rid of that using this dracut
tuning.

References for some analysis about that:

https://gitlab.steamos.cloud/holo/holo/-/merge_requests/107#note_46441
https://gitlab.steamos.cloud/jupiter/tasks/-/issues/640#note_73458

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>