steamos-chroot: Add option to not mount the /etc overlay
If we try to mount an overlayfs multiple times, `mount` will complain
that the overlay is already mounted and it will fail.

With this commit we add an option to skip the overlay mount.
In this way, if we use the new `--no-overlay` option, we will
still be able to create a chroot for "self" and also launch multiple
chroots for the same disk.

Fixes: steam/steamos-customizations#2

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>