steamos-chroot: Allow multiple concurrent chroots into the same partset
Given that we can't mount two times the same overlay, we were not able
to launch two instances of `steamos-chroot` for the same disk and
partset.

To avoid this limitation, we mount the appropriate /etc overlay in
`/run`, and then bind mount it to all `steamos-chroot` instances that
require it.

By using a shared lock, the last chroot that will exit, will also
unmount the /etc overlay.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>