steamos-chroot: Add an explicit --allow-self option
A chroot into the self, or currently active partition, is usually a
synonym of a mistake. E.g. when a user chose A while mistakenly
believing he was in B.

For this reason, we try to catch this situation and fail early with an
informative message.

If instead we really wanted to chroot into the currently active
partition, we can force it by using the option `--allow-self`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>