steamos-factory-reset: rework config file and handling
The current approach suffers a few issues:
 - stores the full feature set, throwing them all away but casefold
   - and does not actually use the casefold :facepalm:
 - hard-codes the (lack of) reserved blocks
 - uses bash arrays

Address those, fetching and storing only the data we need. Use the same
code in both config generator and parser.

v2
 - split other fixes in prep patches
 - restore the old per-partition config

The separate configs ensure that if the formatting fails, a follow-up
boot will need to re-trigger again. Currently we boot _even_ if the
format fails, which effectively avoids a boot loop scenario, yet it's
great user experience. Fixing that requires a complete redesign.

v3
 - fix typo in the config generator

v4
 - drop the /proc/mount code and config handling

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>