steamos-factory-reset: remove unreachable (u)mount code
During the early boot stage, no (user) partitions are mounted so this
code is unreachable. Most likely it is a remnant of the old factory
reset, which was happening during poweroff/reboot and the partition(s)
might have still been mounted.

Remove it since it is dead (and actually slightly buggy) code.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>