steamos-factory-reset: remove misleading readlink code
In the reset_device_ext4() the device passed is guaranteed to
non-symlink, as seen in the reset-config script.

Whereas for factory_reset(), we don't need the readlink code since mount
and ismount work correctly with the /dev/disks/by-partsets/XXX symlinks.

We extensively rely on that in the rootfs mounting code et al.

With those gone, we can drop the (explicit) readlink from the initrd.
Chances are other components will pull it in, but that's beyond our
control.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>