dracut: remove unused tools from initrd
With the earlier factory-reset rework a bunch of tools got added into
the initrd, while they are unused. Seemingly no user is on the horizon,
so let's remove them and reduce the bloat.

Glancing through the list - a few others (like sort) can also go
although that can happen at a later stage.

Fixes: ee43610 ("Implement factory-reset-from bootloader via kernel command line")
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>