mkinitcpio: add pacman hook, managing the conf
Change the existing config to include the required hooks in the specific
order. Upstream has WIP to handle that as a drop-in file, but for now we
need the hook and script.

Earlier versions of this patch were also managing the preset but that
was since been removed. It was fragile and quite extra complexity for
no meaningful reason - the extra space/couple of seconds for the
fallback initrd are negligible.

v2
 - s/present/preset/g

v3
 - use official hook.preset as source

v4
 - drop all preset handling

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>