steamos-factory-reset-config: don't take an optional dir
The tool already knows where the reset config should be generated. There
is no documentation (formal or a help screen) nor shell completion, to
indicate that it takes an optional argument, which is the config
directory.

So as-is you get nicely surpriced if you provide another directory and
the reset does not work. Using a simple `--help` arg also blows up.

Remove the argument handling and simplify all the related code.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>