Fix the preservation of ext4 filesystem features
Filesystem features were being dropped in favour of the defaults
from mkfs.ext4: They are now copied from the old to new fs.

Reserved block percentage is set to 0 if the reserved count was 0 on
the old filesystem instead of tying this to the filesystem label being
"home".

If the reserved count was > 0 then we let mkfs set the default
reserved percentage for the new fs.